requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
email-validator==2.1.0
WTForms==3.1.1